Fix file:/ URLs on Windows
A URL like `file:/C:\foo\bar\baz.jpg` was trying to read
`/C:\foo\bar\baz.jpg`. This change strips the leading slash if we're on
windows and the path looks like it starts with a drive letter. It also
uses `filepath.FromSlash` to ensure that the path has the correct
separator (`/` or `\`) throughout.